#React Components implementation as per design
